+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # Changelog
 
+## v0.2.0 - [2024-06-13]
+* Rewrite substitution logic from string replace to re.sub.
+* Add Template.DELIMITER classvar.
+* Add `benchmarks` make-command.
+* Actualize benchmarks and tests.
+* Update README.
+
 ## v0.1.0 - [2024-06-08]
 * First release with main functionality.